requirements.txt
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.21
flake8==3.8.4
Flask==1.1.2
hdbscan==0.8.26
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mysql-connector-python==8.0.22
numpy==1.19.3
pandas==1.1.4
Pillow==8.0.1
protobuf==3.14.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pylama==7.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
